# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "0.5.0",
        "title": "Projeto Integrador SmartWater API",
        "description": "API criada a partir da biblioteca da Express, com o uso da linguagem Typescript, a API foi criada para obter dados de Caixas d'águas, Hidromêtros e Artesians emitidos por sensores",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/license/MIT"
        }
    },
    "host": "localhost:3000",
    "schemes": ["http", "https"],
    "basePath": "/api/timeseries/v0.5/smartcampusmaua/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
...
    "paths": {
...
        "/Hidrometers": {
            "get" : {
                "tags": [
                    "Hidrometers"
                ],
                "parameters": [
                    {
                        "name": "interval",
                        "in": "query",
                        "type": "number",
                        "description": "Query parameter that sets the interval of time in Minutes to get the data",
                        "default": "15",
                        "required": false
                    }
                ],
                "summary": "Get All Data emited by the six sensors for the Hidrometers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "example": {
                                "Hidrometer_1": {
                                    "data_boardVoltage": [
                                        {
                                            "fieldValue": 1.126,
                                            "timestamp": "2024-05-16T10:46:02.528679842Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z"
                                        },
                                        {
                                            "fieldValue": 1.477,
                                            "timestamp": "2024-05-16T10:49:02.528679842Z",
                                            "start": "2024-05-16T10:22:44.901203384Z",
                                            "stop": "2024-05-16T22:22:44.901203384Z"
                                        }
                                    ],
                                    "data_counter": [
                                        {
                                            "fieldValue": 8,
                                            "timestamp": "2024-05-16T10:25:09.853963392Z",
                                            "start": "2024-05-16T10:22:41.901203384Z",
                                            "stop": "2024-05-16T22:22:41.901203384Z"
                                        },
                                        {
                                            "fieldValue": 24,
                                            "timestamp": "2024-05-16T10:25:16.850516604Z",
                                            "start": "2024-05-16T10:22:41.901203384Z",
                                            "stop": "2024-05-16T22:22:42.901203384Z"
                                        }
                                    ],
                                    "...": ["..."]
                                },
                                "..." : {"..." : [{"...": "..."}]},
                                "Hidrometer_6": {
                                    "data_boardVoltage": [
                                        {
                                            "fieldValue": 2.126,
                                            "timestamp": "2024-05-16T10:46:02.528679842Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        },
                                        {
                                            "fieldValue": 1.834,
                                            "timestamp": "2024-05-16T10:49:02.528679842Z",
                                            "start": "2024-05-16T10:22:44.901203384Z",
                                            "stop": "2024-05-16T22:22:44.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "data_counter": [
                                        {
                                            "fieldValue": 16,
                                            "timestamp": "2024-05-16T10:25:09.853963392Z",
                                            "start": "2024-05-16T10:22:41.901203384Z",
                                            "stop": "2024-05-16T22:22:41.901203384Z",
                                            "devEUI": 8066
                                        },
                                        {
                                            "fieldValue": 27,
                                            "timestamp": "2024-05-16T10:25:16.850516604Z",
                                            "start": "2024-05-16T10:22:41.901203384Z",
                                            "stop": "2024-05-16T22:22:42.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/Hidrometers/deviceName/{nodeName}": {
            "get" : {
                "tags": [
                    "Hidrometers"
                ],
                "parameters": [
                    {
                        "name": "nodeName",
                        "in": "path",
                        "type": "string",
                        "description": "Param that receives the name of the Hidrometer node, like Hidrometer_1",
                        "required": true
                    }
                ],
                "summary": "Get All Data emited by the Eight sensors installed for the Hidrometers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "example": {
                                "Hidrometer_1": {
                                    "data_boardVoltage": [
                                        {
                                            "fieldValue": 1.126,
                                            "timestamp": "2024-05-16T10:46:02.528679842Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z"
                                        },
                                        {
                                            "fieldValue": 1.477,
                                            "timestamp": "2024-05-16T10:49:02.528679842Z",
                                            "start": "2024-05-16T10:22:44.901203384Z",
                                            "stop": "2024-05-16T22:22:44.901203384Z"
                                        }
                                    ],
                                    "data_counter": [
                                        {
                                            "fieldValue": 8,
                                            "timestamp": "2024-05-16T10:25:09.853963392Z",
                                            "start": "2024-05-16T10:22:41.901203384Z",
                                            "stop": "2024-05-16T22:22:41.901203384Z",
                                            "devEUI": 8066
                                        },
                                        {
                                            "fieldValue": 24,
                                            "timestamp": "2024-05-16T10:25:16.850516604Z",
                                            "start": "2024-05-16T10:22:41.901203384Z",
                                            "stop": "2024-05-16T22:22:42.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "...": ["..."]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "message": "The Route param nodeName is missing!"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "example": {
                                "message": "There is not a value with Nodename that you have passed!"
                            }
                        }
                    }
                }
            }
        },
        "/Hidrometers/deviceId/{devEUI}": {
            "get" : {
                "tags": [
                    "Hidrometers"
                ],
                "parameters": [
                    {
                        "name": "devEUI",
                        "in": "path",
                        "type": "string",
                        "description": "Param that receives the Id of the Hidrometer node, 2148",
                        "required": true
                    }
                ],
                "summary": "Get Data emited filtering with the device ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "example": {
                                "Hidrometer_1": {
                                    "data_boardVoltage": [
                                        {
                                            "fieldValue": 1.126,
                                            "timestamp": "2024-05-16T10:46:02.528679842Z",
                                            "start": "2024-05-16T10:22:43.901203384Z",
                                            "stop": "2024-05-16T22:22:43.901203384Z",
                                            "devEUI": 8066
                                        },
                                        {
                                            "fieldValue": 1.477,
                                            "timestamp": "2024-05-16T10:49:02.528679842Z",
                                            "start": "2024-05-16T10:22:44.901203384Z",
                                            "stop": "2024-05-16T22:22:44.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "data_counter": [
                                        {
                                            "fieldValue": 8,
                                            "timestamp": "2024-05-16T10:25:09.853963392Z",
                                            "start": "2024-05-16T10:22:41.901203384Z",
                                            "stop": "2024-05-16T22:22:41.901203384Z",
                                            "devEUI": 8066
                                        },
                                        {
                                            "fieldValue": 24,
                                            "timestamp": "2024-05-16T10:25:16.850516604Z",
                                            "start": "2024-05-16T10:22:41.901203384Z",
                                            "stop": "2024-05-16T22:22:42.901203384Z",
                                            "devEUI": 8066
                                        }
                                    ],
                                    "...": ["..."]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "example": {
                                "message": "The Route param devEUI is missing!"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "example": {
                                "message": "There is not a value with devEUI that you have passed!"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}